v0.9/pyproject.toml
[project]
name = "llm-css"
version = "0.1.0"
description = ""
authors = [
    {name = "Zhoukun Zhou",email = "zhenkun91@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.11, <3.15"
dependencies = [
    "tweet-preprocessor (>=0.6.0,<0.7.0)",
    "scikit-learn (>=1.7.2,<2.0.0)",
    "torch (>=2.9.0,<3.0.0)",
    "transformers (>=4.57.1,<5.0.0)",
    "wordninja (>=2.0.0,<3.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "pendulum (>=3.1.0,<4.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "modelscope (>=1.31.0,<2.0.0)",
    "addict (>=2.4.0,<3.0.0)",
    "emoji (>=2.15.0,<3.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
priority = "primary"
